refactor: move `String.toString` -> `Objects.toString`
