+++ lib/functions/Strings.js
@@ -27,21 +27,6 @@ function templater(template, context = {}) {
 }
 
 /**
- * To string
- * @param {any} o - the object to get it text representation
- * @returns {string} the `o` as string
- */
-function toString(o) {
-  // null or undefined
-  if (o === null || o === void 0) return o;
-  if (typeof o === "string") return o;
-  // has a toString function in their prototype
-  if (typeof o.toString === "function") return o.toString();
-  // as string in the latest intent
-  return String(o);
-}
-
-/**
  * Wraps a string between other that acts as token.
  * @param {string} s - the text to wrap
  * @param {string} token - the text to wrap with between
@@ -56,6 +41,5 @@ function wrap(s, token = "") {
 module.exports = {
   stripParens,
   templater,
-  toString,
   wrap,
 };
