+++ lib/functions/index.js
@@ -1,5 +1,7 @@
+const Objects = require("./Objects");
 const Strings = require("./Strings");
 
 module.exports = {
+  Objects,
   Strings,
 };
